-define custom user registration form (for use in django-admin) in forms.py as opposed to admin.py
