+++ admin/common_auth/admin.py
@@ -1,22 +1,12 @@
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin
 from django.contrib.auth.models import Permission
+from django.contrib.auth.forms import PasswordResetForm
 from .models import MyUser
-
-from django.contrib.auth.forms import PasswordResetForm, UserCreationForm
+from forms import CustomUserRegistrationForm
 
 class PermissionAdmin(admin.ModelAdmin):
     search_fields = ['name', 'codename']
-
-class CustomUserRegistrationForm(UserCreationForm):
-    class Meta:
-            model = MyUser
-            fields = ['password', 'first_name', 'last_name', 'email', 'is_active', 'is_staff',
-            'is_superuser', 'groups', 'user_permissions', 'last_login', ]
-    def __init__(self, *args, **kwargs):
-        super(CustomUserRegistrationForm, self).__init__(*args, **kwargs)
-        self.fields['first_name'].required = True
-        self.fields['last_name'].required = True
 
 class CustomUserAdmin(UserAdmin):
     add_form = CustomUserRegistrationForm
